almost perfect condition for deployment

+++ build_deploy_scripts/artifact_api.py
@@ -12,7 +12,7 @@ class ArtifactApi(object):
         self.registry_base_path = "/mnt/eed/ad-build/registry/"
         self.artifact_api_url = "http://artifact-api-service.core-build-system:8080/"
 
-    def download_file_response(self, download_dir: str, tag: str, response: requests.Response):
+    def download_file_response(self, download_dir: str, tag: str, response: requests.Response, extract: bool):
         # Download file from api, and extract to download_dir
         # Download the .tar.gz file
         tarball_filepath = download_dir + '/' + tag + '.tar.gz'
@@ -25,20 +25,21 @@ class ArtifactApi(object):
                         file.write(chunk)
             logger.info('Tarball downloaded successfully')
             # Extract the .tar.gz file
-            logger.info('Extracting tarball...')
-            with tarfile.open(tarball_filepath, 'r:gz') as tar:
-                tar.extractall(path=download_dir)
-            logger.info(f'{tarball_filepath} extracted to {download_dir}')
-            # Delete tarball after extracting
-            os.remove(tarball_filepath)
+            if (extract):
+                logger.info('Extracting tarball...')
+                with tarfile.open(tarball_filepath, 'r:gz') as tar:
+                    tar.extractall(path=download_dir)
+                logger.info(f'{tarball_filepath} extracted to {download_dir}')
+                # Delete tarball after extracting
+                os.remove(tarball_filepath)
         else:
             logger.info('Failed to retrieve the file. Status code:', response.status_code)
 
 
-    def get_component_from_registry(self, download_dir: str, component: str, tag: str, os_env: str = None):   
+    def get_component_from_registry(self, download_dir: str, component: str, tag: str, os_env: str = 'null', extract: bool = True):   
         payload = {"component": component, "tag": tag, "arch": os_env}
         logger.info(f"Get component {component},{tag} request to artifact storage...")
         # stream=True in case it's a large tarball
         response = requests.get(url=self.artifact_api_url + 'component', json=payload, stream=True) 
-        self.download_file_response(download_dir, tag, response)
+        self.download_file_response(download_dir, tag, response, extract)
         # For now we can assume the component exists, otherwise the api builds and returns it
